API action result is shown after the call is made

+++ tfg_frontend/lib/endpoints/Objects/BuildingDemandDispersions.dart
@@ -3,7 +3,7 @@ import 'dart:ffi';
 class BuildingDemand {
   final int building_type;
   final String climatic_zone;
-  final Float dispersion;
+  final double dispersion;
 
   const BuildingDemand({
     required this.building_type,
